Move setup_config_file fixtures to conftest
