+++ tests/conftest.py
@@ -134,3 +134,14 @@ def setup_test_dir_paths(
     ]
 
     return paths
+
+
+@pytest.fixture
+def setup_config_file(setup_env) -> Path:
+    """
+    Fixture to set up a temporary config file.
+    """
+    config_dir = setup_env / "config"
+    config_dir.mkdir(exist_ok=True)
+    config_file = config_dir / "config.json"
+    return config_file
